Split mailbox handling out into a service
